feat: add category pages
