+++ src/app/(app)/(home)/[category]/[subcategory]/page.tsx
@@ -1,0 +1,14 @@
+interface Props {
+	params: Promise<{
+		category: string;
+		subcategory: string;
+	}>;
+}
+
+const Page = async ({ params }: Props) => {
+	const { category, subcategory } = await params;
+
+	return <div>Subcategory Page: {subcategory} in Category: {category}</div>;
+};
+
+export default Page;
