+++ src/app/(app)/(home)/[category]/page.tsx
@@ -1,0 +1,13 @@
+interface Props {
+	params: Promise<{
+		category: string;
+	}>;
+}
+
+const Page = async ({ params }: Props) => {
+	const { category } = await params;
+
+	return <div>Category Page: {category}</div>;
+};
+
+export default Page;
