+++ src/app/(app)/(home)/layout.tsx
@@ -3,9 +3,9 @@ import { dehydrate, HydrationBoundary } from '@tanstack/react-query';
 
 import { getQueryClient, trpc } from '@/trpc/server';
 
-import Navbar from './navbar';
-import Footer from './footer';
-import SearchFilters, { SearchFiltersSkeleton } from './search-filters';
+import Navbar from '@/modules/home/ui/components/navbar';
+import Footer from '@/modules/home/ui/components/footer';
+import SearchFilters, { SearchFiltersSkeleton } from '@/modules/home/ui/components/search-filters';
 
 interface LayoutProps {
 	children: ReactNode;
