+++ src/modules/home/ui/components/constants.ts
@@ -1,0 +1,1 @@
+export const DEFAULT_BG_COLOR = '#F5F5F5';
