+++ src/modules/home/ui/components/footer.tsx
@@ -1,0 +1,11 @@
+const Footer = () => {
+  return (
+    <footer className="flex border-t justify-between font-medium p-6">
+      <div className="flex items-center gap-2">
+        <p>funroad, Inc.</p>
+      </div>
+    </footer>
+  )
+}
+
+export default Footer
